+++ config.c
@@ -339,7 +339,7 @@ int parse(char *buf, int linenum)
     dir = strtok(buf, " \t\r\n");
     s = strtok(NULL, "");
     if (s) {
-	while (isspace(*s))
+	while (isspace((int)*s))
 	    s++;
 	while (*s) {
 	    if (ac >= MAXPARAMS) {
@@ -366,7 +366,7 @@ int parse(char *buf, int linenum)
 		    *s++ = 0;
 	    }
 	    av[ac++] = t;
-	    while (isspace(*s))
+	    while (isspace((int)*s))
 		s++;
 	}
     }
